requirements.txt
python-twitter==3.4.1
django==1.11